backend/requirements.txt
fastapi==0.115.6
uvicorn==0.25.0
python-multipart==0.0.9
langchain==0.3.11
langchain-huggingface==0.1.2
langchain-weaviate==0.0.3
langchain-openai==0.2.11
langchain-openai
FlashRank==0.2.9
pydantic==2.9.2
weaviate-client==4.9.6
pymongo==4.10.1
pandas==1.5.3